+++ ipwhois/asn.py
@@ -25,7 +25,8 @@
 import re
 import copy
 import logging
-from . import (NetError)
+from .exceptions import (NetError, ASNRegistryError, ASNParseError,
+                         ASNLookupError, HTTPLookupError)
 
 log = logging.getLogger(__name__)
 
@@ -37,7 +38,7 @@ BASE_NET = {
     'source': None
 }
 
-ASN_WHOIS = {
+ASN_ORIGIN_WHOIS = {
     'radb': {
         'server': 'whois.radb.net',
         'fields': {
@@ -49,7 +50,7 @@ ASN_WHOIS = {
     },
 }
 
-ASN_HTTP = {
+ASN_ORIGIN_HTTP = {
     'radb': {
         'url': 'http://www.radb.net/query/{0}',
         'fields': {
@@ -60,6 +61,326 @@ ASN_HTTP = {
         }
     },
 }
+
+
+class IPASN:
+    """
+    The class for parsing ASN data for an IP address.
+
+    Args:
+        net: A ipwhois.net.Net object.
+
+    Raises:
+        NetError: The parameter provided is not an instance of
+            ipwhois.net.Net
+    """
+
+    def __init__(self, net):
+
+        from .net import (Net, ORG_MAP)
+        from .whois import RIR_WHOIS
+
+        # ipwhois.net.Net validation
+        if isinstance(net, Net):
+
+            self._net = net
+
+        else:
+
+            raise NetError('The provided net parameter is not an instance of '
+                           'ipwhois.net.Net')
+
+        self.org_map = ORG_MAP
+        self.rir_whois = RIR_WHOIS
+
+    def _parse_fields_dns(self, response):
+        """
+        The function for parsing ASN fields from a dns response.
+
+        Args:
+            response: The response from the ASN dns server.
+
+        Returns:
+            Dictionary:
+
+            :asn: The Autonomous System Number (String)
+            :asn_date: The ASN Allocation date (String)
+            :asn_registry: The assigned ASN registry (String)
+            :asn_cidr: The assigned ASN CIDR (String)
+            :asn_country_code: The assigned ASN country code (String)
+
+        Raises:
+            ASNRegistryError: The ASN registry is not known.
+            ASNParseError: ASN parsing failed.
+        """
+
+        try:
+
+            temp = response.split('|')
+
+            # Parse out the ASN information.
+            ret = {'asn_registry': temp[3].strip(' \n')}
+
+            if ret['asn_registry'] not in self.rir_whois.keys():
+
+                raise ASNRegistryError(
+                    'ASN registry {0} is not known.'.format(
+                        ret['asn_registry'])
+                )
+
+            ret['asn'] = temp[0].strip(' "\n')
+            ret['asn_cidr'] = temp[1].strip(' \n')
+            ret['asn_country_code'] = temp[2].strip(' \n').upper()
+            ret['asn_date'] = temp[4].strip(' "\n')
+
+        except ASNRegistryError:
+
+            raise
+
+        except Exception as e:
+
+            raise ASNParseError('Parsing failed for "{0}" with exception: {1}.'
+                                ''.format(response, e)[:100])
+
+        return ret
+
+    def _parse_fields_whois(self, response):
+        """
+        The function for parsing ASN fields from a whois response.
+
+        Args:
+            response: The response from the ASN whois server.
+
+        Returns:
+            Dictionary:
+
+            :asn: The Autonomous System Number (String)
+            :asn_date: The ASN Allocation date (String)
+            :asn_registry: The assigned ASN registry (String)
+            :asn_cidr: The assigned ASN CIDR (String)
+            :asn_country_code: The assigned ASN country code (String)
+
+        Raises:
+            ASNRegistryError: The ASN registry is not known.
+            ASNParseError: ASN parsing failed.
+        """
+
+        try:
+
+            temp = response.split('|')
+
+            # Parse out the ASN information.
+            ret = {'asn_registry': temp[4].strip(' \n')}
+
+            if ret['asn_registry'] not in self.rir_whois.keys():
+
+                raise ASNRegistryError(
+                    'ASN registry {0} is not known.'.format(
+                        ret['asn_registry'])
+                )
+
+            ret['asn'] = temp[0].strip(' \n')
+            ret['asn_cidr'] = temp[2].strip(' \n')
+            ret['asn_country_code'] = temp[3].strip(' \n').upper()
+            ret['asn_date'] = temp[5].strip(' \n')
+
+        except ASNRegistryError:
+
+            raise
+
+        except Exception as e:
+
+            raise ASNParseError('Parsing failed for "{0}" with exception: {1}.'
+                                ''.format(response, e)[:100])
+
+        return ret
+
+    def _parse_fields_http(self, response, extra_org_map=None):
+        """
+        The function for parsing ASN fields from a http response.
+
+        Args:
+            response: The response from the ASN http server.
+            extra_org_map: Dictionary mapping org handles to RIRs. This is for
+                limited cases where ARIN REST (ASN fallback HTTP lookup) does
+                not show an RIR as the org handle e.g., DNIC (which is now the
+                built in ORG_MAP) e.g., {'DNIC': 'arin'}. Valid RIR values are
+                (note the case-sensitive - this is meant to match the REST
+                result): 'ARIN', 'RIPE', 'apnic', 'lacnic', 'afrinic'
+
+        Returns:
+            Dictionary:
+
+            :asn: None, can't retrieve with this method.
+            :asn_date: None, can't retrieve with this method.
+            :asn_registry: The assigned ASN registry (String)
+            :asn_cidr: None, can't retrieve with this method.
+            :asn_country_code: None, can't retrieve with this method.
+
+        Raises:
+            ASNRegistryError: The ASN registry is not known.
+            ASNParseError: ASN parsing failed.
+        """
+
+        # Set the org_map. Map the orgRef handle to an RIR.
+        org_map = self.org_map.copy()
+        try:
+
+            org_map.update(extra_org_map)
+
+        except (TypeError, ValueError, IndexError, KeyError):
+
+            pass
+
+        try:
+
+            asn_data = {
+                'asn_registry': None,
+                'asn': None,
+                'asn_cidr': None,
+                'asn_country_code': None,
+                'asn_date': None
+            }
+
+            try:
+
+                net_list = response['nets']['net']
+
+                if not isinstance(net_list, list):
+                    net_list = [net_list]
+
+            except (KeyError, TypeError):
+
+                log.debug('No networks found')
+                net_list = []
+
+            for n in net_list:
+
+                try:
+
+                    asn_data['asn_registry'] = (
+                        org_map[n['orgRef']['@handle'].upper()]
+                    )
+
+                except KeyError as e:
+
+                    log.debug('Could not parse ASN registry via HTTP: '
+                              '{0}'.format(str(e)))
+                    raise ASNRegistryError('ASN registry lookup failed.')
+
+                break
+
+        except ASNRegistryError:
+
+            raise
+
+        except Exception as e:  # pragma: no cover
+
+            raise ASNParseError('Parsing failed for "{0}" with exception: {1}.'
+                                ''.format(response, e)[:100])
+
+        return asn_data
+
+    def lookup(self, inc_raw=False, retry_count=3, asn_alts=None,
+               extra_org_map=None):
+        """
+        The wrapper function for retrieving and parsing ASN information for an
+        IP address.
+
+        Args:
+            inc_raw: Boolean for whether to include the raw results in the
+                returned dictionary.
+            retry_count: The number of times to retry in case socket errors,
+                timeouts, connection resets, etc. are encountered.
+            asn_alts: Array of additional lookup types to attempt if the
+                ASN dns lookup fails. Allow permutations must be enabled.
+                Defaults to all ['whois', 'http'].
+            extra_org_map: Dictionary mapping org handles to RIRs. This is for
+                limited cases where ARIN REST (ASN fallback HTTP lookup) does
+                not show an RIR as the org handle e.g., DNIC (which is now the
+                built in ORG_MAP) e.g., {'DNIC': 'arin'}. Valid RIR values are
+                (note the case-sensitive - this is meant to match the REST
+                result): 'ARIN', 'RIPE', 'apnic', 'lacnic', 'afrinic'
+
+        Returns:
+            Dictionary:
+
+            :asn: The Autonomous System Number (String)
+            :asn_date: The ASN Allocation date (String)
+            :asn_registry: The assigned ASN registry (String)
+            :asn_cidr: The assigned ASN CIDR (String)
+            :asn_country_code: The assigned ASN country code (String)
+            :raw: Raw ASN results if the inc_raw parameter is True. (String)
+
+        Raises:
+            ASNRegistryError: ASN registry does not match.
+            HTTPLookupError: The HTTP lookup failed.
+        """
+
+        lookups = asn_alts if asn_alts is not None else ['whois', 'http']
+
+        # Attempt to resolve ASN info via Cymru. DNS is faster, try that first.
+        try:
+
+            self._net.dns_resolver.lifetime = (
+                self._net.dns_resolver.timeout * (
+                    retry_count and retry_count or 1
+                )
+            )
+            response = self._net.get_asn_dns()
+            asn_data = self._parse_fields_dns(response)
+
+        except (ASNLookupError, ASNRegistryError) as e:
+
+            if not self._net.allow_permutations:
+
+                raise ASNRegistryError('ASN registry lookup failed. '
+                                       'Permutations not allowed.')
+
+            try:
+                if 'whois' in lookups:
+
+                    log.debug('ASN DNS lookup failed, trying ASN WHOIS: '
+                              '{0}'.format(e))
+                    response = self._net.get_asn_whois(retry_count)
+                    asn_data = self._parse_fields_whois(response)
+
+                else:
+
+                    raise ASNLookupError
+
+            except (ASNLookupError, ASNRegistryError):  # pragma: no cover
+
+                if 'http' in lookups:
+
+                    # Lets attempt to get the ASN registry information from
+                    # ARIN.
+                    log.debug('ASN WHOIS lookup failed, trying ASN via HTTP')
+                    try:
+
+                        response = self._net.get_asn_http(
+                            retry_count=retry_count
+                        )
+                        asn_data = self._parse_fields_http(response,
+                                                           extra_org_map)
+
+                    except ASNRegistryError:
+
+                        raise ASNRegistryError('ASN registry lookup failed.')
+
+                    except ASNLookupError:
+
+                        raise HTTPLookupError('ASN HTTP lookup failed.')
+
+                else:
+
+                    raise ASNRegistryError('ASN registry lookup failed.')
+
+        if inc_raw:
+
+            asn_data['raw'] = response
+
+        return asn_data
 
 
 class ASNOrigin:
@@ -227,7 +548,7 @@ class ASNOrigin:
 
             :query: The Autonomous System Number (String)
             :nets: Dictionaries containing network information which consists
-                of the fields listed in the ASN_WHOIS dictionary. (List)
+                of the fields listed in the ASN_ORIGIN_WHOIS dictionary. (List)
             :raw: Raw ASN origin whois results if the inc_raw parameter is
                 True. (String)
         """
@@ -277,7 +598,7 @@ class ASNOrigin:
 
             temp_net = self._parse_fields(
                 response,
-                ASN_WHOIS['radb']['fields'],
+                ASN_ORIGIN_WHOIS['radb']['fields'],
                 section_end,
                 net['end'],
                 field_list