+++ ipwhois/tests/test_asn.py
@@ -3,13 +3,121 @@ import io
 from os import path
 import logging
 from ipwhois.tests import TestCommon
+from ipwhois.exceptions import (ASNRegistryError, ASNLookupError,
+                                ASNParseError)
 from ipwhois.net import Net
-from ipwhois.asn import (ASNOrigin, ASN_WHOIS, ASN_HTTP, NetError)
+from ipwhois.asn import (IPASN, ASNOrigin, ASN_ORIGIN_WHOIS, ASN_ORIGIN_HTTP,
+                         NetError)
 
 LOG_FORMAT = ('[%(asctime)s] [%(levelname)s] [%(filename)s:%(lineno)s] '
               '[%(funcName)s()] %(message)s')
 logging.basicConfig(level=logging.DEBUG, format=LOG_FORMAT)
 log = logging.getLogger(__name__)
+
+
+class TestIPASN(TestCommon):
+
+    def test__IPASN(self):
+
+        self.assertRaises(NetError, IPASN, 'a')
+
+    def test__parse_fields_dns(self):
+
+        data = '"15169 | 74.125.225.0/24 | US | arin | 2007-03-13"'
+        net = Net('74.125.225.229')
+        ipasn = IPASN(net)
+        try:
+            self.assertIsInstance(ipasn._parse_fields_dns(data), dict)
+        except AssertionError as e:
+            raise e
+        except Exception as e:
+            self.fail('Unexpected exception raised: {0}'.format(e))
+
+        data = '"15169 | 74.125.225.0/24 | US | random | 2007-03-13"'
+        self.assertRaises(ASNRegistryError, ipasn._parse_fields_dns, data)
+
+        data = ''
+        self.assertRaises(ASNParseError, ipasn._parse_fields_dns, data)
+
+    def test__parse_fields_whois(self):
+
+        data = ('15169   | 74.125.225.229   | 74.125.225.0/24     | US | arin'
+                '     | 2007-03-13')
+        net = Net('74.125.225.229')
+        ipasn = IPASN(net)
+        try:
+            self.assertIsInstance(ipasn._parse_fields_whois(data), dict)
+        except AssertionError as e:
+            raise e
+        except Exception as e:
+            self.fail('Unexpected exception raised: {0}'.format(e))
+
+        data = ('15169   | 74.125.225.229   | 74.125.225.0/24     | US | rdm'
+                '     | 2007-03-13')
+        self.assertRaises(ASNRegistryError, ipasn._parse_fields_whois, data)
+
+        data = '15169   | 74.125.225.229   | 74.125.225.0/24     | US'
+        self.assertRaises(ASNParseError, ipasn._parse_fields_whois, data)
+
+    def test__parse_fields_http(self):
+
+        data = {
+            'nets': {
+                'net': {
+                    'orgRef': {
+                        '@handle': 'APNIC'
+                    }
+                }
+            }
+        }
+        net = Net('1.2.3.4')
+        ipasn = IPASN(net)
+        try:
+            self.assertIsInstance(ipasn._parse_fields_http(response=data),
+                                  dict)
+        except AssertionError as e:
+            raise e
+        except Exception as e:
+            self.fail('Unexpected exception raised: {0}'.format(e))
+
+        data['nets']['net']['orgRef']['@handle'] = 'RIPE'
+        try:
+            self.assertIsInstance(ipasn._parse_fields_http(response=data),
+                                  dict)
+        except AssertionError as e:
+            raise e
+        except Exception as e:
+            self.fail('Unexpected exception raised: {0}'.format(e))
+
+        data['nets']['net']['orgRef']['@handle'] = 'DNIC'
+        try:
+            self.assertIsInstance(ipasn._parse_fields_http(response=data),
+                                  dict)
+        except AssertionError as e:
+            raise e
+        except Exception as e:
+            self.fail('Unexpected exception raised: {0}'.format(e))
+
+        data['nets']['net']['orgRef']['@handle'] = 'INVALID'
+        try:
+            self.assertRaises(ASNRegistryError, ipasn._parse_fields_http,
+                              response=data)
+        except AssertionError as e:
+            raise e
+        except Exception as e:
+            self.fail('Unexpected exception raised: {0}'.format(e))
+
+        data = ''
+        try:
+            self.assertIsInstance(ipasn._parse_fields_http(response=data), dict)
+        except AssertionError as e:
+            raise e
+        except Exception as e:
+            self.fail('Unexpected exception raised: {0}'.format(e))
+
+    def test__IPASNLookup(self):
+        # TODO: need to modify asn.json for this.
+        return NotImplemented
 
 
 class TestASNOrigin(TestCommon):
@@ -54,7 +162,7 @@ class TestASNOrigin(TestCommon):
 
         # No exception raised, but should provide code coverage for if regex
         # groups are messed up.
-        tmp_dict = ASN_WHOIS['radb']['fields']
+        tmp_dict = ASN_ORIGIN_WHOIS['radb']['fields']
         tmp_dict['route'] = r'(route):[^\S\n]+(?P<val1>.+?)\n'
         obj._parse_fields(
             response="\nroute:        66.249.64.0/20\n",
@@ -63,7 +171,7 @@ class TestASNOrigin(TestCommon):
 
         obj._parse_fields(
             response="\nchanged:        noc@google.com 20110301\n",
-            fields_dict=ASN_WHOIS['radb']['fields']
+            fields_dict=ASN_ORIGIN_WHOIS['radb']['fields']
         )
 
         multi_net_response = (
@@ -85,7 +193,7 @@ class TestASNOrigin(TestCommon):
         )
         obj._parse_fields(
             response=multi_net_response,
-            fields_dict=ASN_WHOIS['radb']['fields']
+            fields_dict=ASN_ORIGIN_WHOIS['radb']['fields']
         )
 
     def test__get_nets_radb(self):